+++ src/components/TechniqueChips.vue
@@ -11,7 +11,7 @@
       v-for="tech in techniques.filter(t => t !== '')"
       :key="tech"
       small
-      :href="`https://attack.mitre.org/techniques/${tech}`"
+      :href="getTechniqueUrl(tech)"
       target="_blank"
       color="green"
       class="mr-1 mb-1"
@@ -32,6 +32,15 @@ export default {
     showTitle: {
       type: Boolean,
       default: true,
+    },
+  },
+  methods: {
+    getTechniqueUrl(tech) {
+      if (tech.includes('.')) {
+        const [main, sub] = tech.split('.');
+        return `https://attack.mitre.org/techniques/${main}/${sub}`;
+      }
+      return `https://attack.mitre.org/techniques/${tech}`;
     },
   },
 };